# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UNIGE Events API
  version: 1.0.0
  description: |
    API REST du backend UNIGE Events (Java 21 / Quarkus 3.32).
    Source de vérité : unige-events-api/docs/openapi/openapi.yaml
    Ce fichier est synchronisé avec le repo frontend. Mettre à jour les deux repos ensemble.
    Tous les champs sont en camelCase. Les booléens n'ont pas de préfixe is.
servers:
  - url: /api
    description: Proxié par Nginx (prod) ou Vite (dev)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT   # Auth0 / OIDC — validé par quarkus-oidc
...
  schemas:
...
    Faculty:
      type: string
      enum: [SCIENCES, LETTRES, DROIT, MEDECINE, SES, PSYCHOLOGIE, THEOLOGIE, FTI, GSI]
...
    EventCategory:
      type: string
      description: Catégorie d'un événement
      enum: [ACADEMIC, SPORTS, CULTURAL, SOCIAL, CONFERENCE, OTHER]
    EventStatus:
      type: string
      description: |
        Statut d'un événement.

        - `DRAFT`/`PUBLISHED`/`CANCELLED` sont définis par l'organisateur (un
          `CANCELLED` reste réversible vers `DRAFT` via `PATCH /events/{id}/restore`).
        - `EXPIRED` est attribué automatiquement par le job d'expiration quand un
          PUBLISHED voit sa `endDate` passer dans le passé. Ne peut pas être posé
          via l'API.
        - `BANNED` (SCRUM-97) est attribué par modération — soit manuellement par
          un admin via `PATCH /admin/reports/{id}` (status=REVIEWED), soit
          automatiquement par `ModerationCleanupService` au seuil de reports
          PENDING. État terminal côté créateur : pas de republish possible. Ne
          peut pas être posé via l'API utilisateur.
      enum: [DRAFT, PUBLISHED, CANCELLED, EXPIRED, BANNED]
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
      required: [error, message, details]
    ValidationErrorDetail:
      type: object
      properties:
        field:
          type: string
          nullable: true
        message:
          type: string
      required: [message]
...
paths:
...
  /events:
    get:
      summary: Liste des événements
      description: Retourne les événements avec pagination et filtres optionnels.
      operationId: getEvents
      tags: [events]
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EventStatus'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/EventCategory'
        - name: organizerId
          in: query
          description: |
            Filtre les événements par créateur (UUID). **Contrainte de sécurité** : lorsque
            `organizerId` est présent, seul `status=PUBLISHED` est autorisé. Si `organizerId` est
            passé sans `status`, le filtre implicite `status=PUBLISHED` est appliqué. Toute
            combinaison avec `status=DRAFT` ou `status=CANCELLED` renvoie `400
            organizer_filter_requires_published`. Pour lister ses propres événements dans
            n'importe quel statut, utiliser `GET /users/me/events`.
          schema:
            type: string
            format: uuid
        - name: endDateFrom
          in: query
          description: Filtre les événements dont endDate >= endDateFrom (datetime ISO local, ex. 2026-04-01T16:57:00). Utilisé par la page d'accueil pour masquer les événements terminés.
          schema:
            type: string
            format: date-time
        - name: faculty
          in: query
          description: Filtre les événements par faculté
          schema:
            $ref: '#/components/schemas/Faculty'
        - name: facultyNone
          in: query
          description: |
            Si true, retourne uniquement les événements dont `faculty` est null
            (non rattachés à une faculté précise). Mutuellement exclusif avec le
            paramètre `faculty` — si les deux sont fournis, `facultyNone` a priorité.
          schema:
            type: boolean
        - name: featured
          in: query
          description: Si true, retourne uniquement les événements actuellement mis en avant (featured=true). Utile pour le dashboard admin.
          schema:
            type: boolean
      responses:
        '200':
          description: Liste des événements (array brut)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: |
            Combinaison de paramètres de requête invalide. Codes d'erreur possibles dans `error` :
            - `organizer_filter_requires_published` : `organizerId` combiné avec un statut non-PUBLISHED.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    post:
      summary: Créer un événement
      description: Crée un nouvel événement. Authentification requise — le créateur est l'utilisateur connecté.
      operationId: createEvent
      tags: [events]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Événement créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: |
            Validation échouée. Pour la récurrence (SCRUM-147) :
            - Bean Validation generic : `frequency` manquant, `maxOccurrences` hors [1,52], `startDate` non-future.
            - `error: "recurrence_unbounded"` : `recurrence` présent mais ni `endDate` ni `maxOccurrences` renseignés.
            - `error: "recurrence_end_before_start"` : `recurrence.endDate < startDate.toLocalDate()`.

            Note : un cap dur de 52 occurrences (parent inclus) s'applique côté générateur ;
            si le calcul `endDate × frequency` produit plus de 52 occurrences, la récurrence
            est silencieusement tronquée à 52 (pas d'erreur). La borne client `maxOccurrences > 52`
            est en revanche rejetée par Bean Validation côté DTO (`@Max(52)`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Profil utilisateur introuvable (appeler GET /users/me d'abord)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /events/{id}:
    get:
      summary: Détail d'un événement
      description: |
        Retourne le détail d'un événement.

        **Règle d'autorisation** : seuls les événements `PUBLISHED` sont accessibles
        anonymement. Les événements `DRAFT` ou `CANCELLED` ne sont visibles que par
        leur créateur, **un co-organisateur ACCEPTED** (cf. SCRUM-136), ou un
        administrateur (rôle `ADMIN`). Dans tout autre cas —
        appelant anonyme, appelant authentifié non-créateur non-co-org non-admin, ou
        appelant authentifié dont le profil n'est pas provisionné — le serveur renvoie
        `404 not_found`, envelope identique à celle pour un ID inexistant (évite un
        oracle d'existence — cf. pentest 2026-04-17 findings 4.12 + 4.15).

        **SCRUM-97 — events `BANNED`** : invisibles via cet endpoint pour TOUT LE MONDE,
        anonyme, créateur et admin compris. Le drill-down sur un event banni passe
        exclusivement par la table de modération (`/admin/reports`). La réponse 404
        utilise la même envelope que pour un ID inexistant — pas de leak du fait que
        l'event a été banni.
      operationId: getEventById
      tags: [events]
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Détail de l'événement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: |
            Événement introuvable, OU événement en statut `DRAFT` / `CANCELLED`
            demandé par un appelant qui n'en est ni le créateur ni un admin.
            Le corps est identique dans les deux cas (pas d'oracle d'existence).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    put:
      summary: Modifier un événement (créateur ou co-organisateur ACCEPTED) — mise à jour complète
      description: |
        Réservé au créateur **ou à un co-organisateur ACCEPTED** (cf. SCRUM-136 cascade
        `isCreatorOrAcceptedCoOrganizer`). Renvoie 403 pour tout autre appelant authentifié.
      operationId: updateEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Événement mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation échouée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Appelant ni créateur ni co-organisateur ACCEPTED (cf. SCRUM-136 cascade)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflit — l'événement est annulé et ne peut pas être modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      summary: Supprimer définitivement un événement annulé (hard-delete)
      description: |
        Supprime définitivement l'événement de la base. Opération irréversible.
        L'événement doit être en statut CANCELLED — sinon 409. Pour annuler un événement,
        utiliser PATCH /events/{id}/cancel d'abord.

        **Réservé au créateur** (suppression physique d'un CANCELLED — non délégable aux
        co-organisateurs ACCEPTED, cf. SCRUM-136 décision 2 : la cascade ne s'étend pas
        aux actions irréversibles).
      operationId: deleteEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Événement supprimé définitivement
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: |
            Appelant non créateur. Le hard-delete reste strict-creator — non délégable
            aux co-organisateurs ACCEPTED (cf. SCRUM-136 décision 2 : action irréversible
            hors scope du « partage de gestion »).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflit — l'événement n'est pas en statut CANCELLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /events/{id}/occurrences:
    get:
      summary: Lister les occurrences d'un événement parent récurrent (SCRUM-147)
      description: |
        Retourne la liste des événements dont `parentEventId` matche `{id}`, triés par
        `startDate ASC, id ASC`. Si l'event ciblé n'est pas un parent récurrent
        (occurrence elle-même, standalone non-récurrent, ou parent sans occurrences
        encore générées), retourne `200 OK` + `[]` — pas de `404`. Le client lit
        `event.recurrenceRule != null` sur `GET /events/{id}` pour distinguer un parent.

        **Visibilité** : passe par `EventService.getById(id, callerAuth0Id, isAdmin)`.
        Un event invisible (DRAFT non-créateur, CANCELLED non-créateur, BANNED, id
        inconnu) retourne `404 not_found` (anti-oracle ISSUE-92).
      operationId: getEventOccurrences
      tags: [events]
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 52
            default: 52
      responses:
        '200':
          description: Liste des occurrences (vide si pas de parent récurrent)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Pagination invalide (`size > 52`, `page < 0`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Event invisible (DRAFT non-créateur, CANCELLED non-créateur, BANNED, id inconnu)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /events/{id}/cancel:
    patch:
      summary: Annuler un événement (DRAFT/PUBLISHED → CANCELLED)
      description: |
        Passe le statut de l'événement à CANCELLED. L'événement reste en base et
        demeure interrogeable. Réservé au créateur **ou à un co-organisateur ACCEPTED**
        (cf. SCRUM-136 cascade). Retourne 409 si déjà annulé.
      operationId: cancelEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Événement annulé — retourne l'EventDTO avec status CANCELLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Appelant ni créateur ni co-organisateur ACCEPTED (cf. SCRUM-136 cascade)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflit — l'événement est déjà annulé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /events/{id}/restore:
    patch:
      summary: Restaurer un événement annulé (CANCELLED → DRAFT)
      description: |
        Remet un événement annulé au statut DRAFT. Réservé au créateur **ou à un
        co-organisateur ACCEPTED** (cf. SCRUM-136 cascade).
        Retourne 409 si l'événement n'est pas en statut CANCELLED.
      operationId: restoreEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Événement restauré — retourne l'EventDTO avec status DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Appelant ni créateur ni co-organisateur ACCEPTED (cf. SCRUM-136 cascade)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflit — l'événement n'est pas en statut CANCELLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /events/{id}/publish:
    patch:
      summary: Publier un événement (DRAFT → PUBLISHED)
      description: |
        Passe le statut de l'événement de DRAFT à PUBLISHED.
        Accessible à tout utilisateur authentifié — le service vérifie ensuite
        que l'appelant est le créateur de l'événement, **un co-organisateur ACCEPTED**
        (cf. SCRUM-136 cascade), ou un administrateur.
        Retourne 409 si l'événement n'est pas en statut DRAFT (déjà publié ou annulé).
      operationId: publishEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Événement publié — retourne l'EventDTO complet avec status PUBLISHED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Appelant ni créateur, ni co-organisateur ACCEPTED, ni ADMIN (cf. SCRUM-136 cascade)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflit — l'événement n'est pas en statut DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: |
            Échec de validation — l'événement ne remplit pas les conditions requises pour être publié
            (titre, lieu, catégorie, date de début future, date de fin cohérente). Le corps contient
            la liste des erreurs à corriger dans `errors`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: validation_failed
                  errors:
                    type: array
                    items:
                      type: string
                    example:
                      - La date de l'événement doit être dans le futur
                      - Le titre est obligatoire
        '429':
          $ref: '#/components/responses/RateLimited'
  /events/{id}/duplicate:
    post:
      summary: Dupliquer un événement existant (clone DRAFT, dates +7 jours)
      description: |
        Crée une copie de l'événement source avec :
        - `status = DRAFT` (toujours, indépendamment du status source — sauf BANNED, refusé en 403)
        - `title = "Copie de <source.title>"` avec dédup via suffixe `(2)`, `(3)`, …, cap N=100
          (au-delà : `422 duplicate_title_collision`)
        - `startDate` et `endDate` décalées de `+7 jours`
        - `creatorId = <caller.id>` (le duplicateur devient organisateur du clone, cf. SCRUM-99 Décision N)
        - `recurrenceRule`, `parentEventId`, `registrationDeadline`, `shareCode`, `featured`,
          `featuredAt` réinitialisés à `null`/`false`
        - Tags copiés (deep copy) ; aucune cascade sur attendances, favorites, views,
          co-organisateurs, comments, reports, attachments (le clone démarre vierge)

        Réservé au créateur, à un co-organisateur ACCEPTED (cascade SCRUM-136), ou à un
        administrateur. Aucune émission Kafka (DRAFT ≠ PUBLISHED).
      operationId: duplicateEvent
      tags: [events]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'événement source à dupliquer
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Event dupliqué — retourne le clone (status DRAFT, dates +7j)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: |
            Appelant ni créateur, ni co-organisateur ACCEPTED, ni ADMIN ; ou source
            `BANNED` (un event banni ne se duplique pas, même par un admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Événement source introuvable (anti-oracle DRAFT/CANCELLED pour non-créateurs)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: |
            Collision de title persistante — `Copie de X (1)` jusqu'à `Copie de X (100)`
            sont tous déjà pris par cet organisateur. Code applicatif : `duplicate_title_collision`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...